requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
PyMuPDF
pytesseract
python-multipart